+++ model/datapack/__init__.py
@@ -1,0 +1,2 @@
+from . import version6
+version6.initPlugin()